[Engine] - Correct a lot of mistake 	 - Update input with controller

+++ includes/engine/AObj.hh
@@ -31,8 +31,8 @@ public:
 public:
   void				setPos(float x, float y);
   std::pair<float, float>	getPos() const;
-  void				setId(int);
-  int				getId(void) const;
+  void				setId(long int);
+  long int				getId(void) const;
 public:
   AGameObject			*getGameObj() const;
   void				setGameObj(AGameObject *);
@@ -44,7 +44,7 @@ protected:
   entityType	_type;
   float		_x;
   float		_y;
-  int		_id;
+  long int		_id;
   AGameObject	*_gameObj;
 };
